#41. Fix more errors reported by MSVC on CI.

+++ unit/render_pipeline_language/kan/render_pipeline_language/compiler_resolve.c
@@ -1863,7 +1863,7 @@ static inline kan_bool_t resolve_match_signature_at_index (struct rpl_compiler_c
                              "incorrect array dimension %ld size: %ld while %ld is expected",
                              context->log_name, module_name, owner_expression->source_name,
                              (long) owner_expression->source_line, (long) signature_index,
-                             get_expression_call_name_for_logging (owner_expression), array_dimension_index,
+                             get_expression_call_name_for_logging (owner_expression), (long) array_dimension_index,
                              (long) expression->output.type.array_dimensions[array_dimension_index],
                              (long) signature->variable.type.array_dimensions[array_dimension_index])
                     return KAN_FALSE;
